requirements.txt
Babel==2.5.0
bleach==2.0.0
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
cssmin==0.2.0
diff-match-patch==20121119
dirsync==2.2.2
Django==1.10.8
django-allauth==0.33.0
django-appconf==1.0.2
django-assets==0.12
django-bulk-update==2.2.0
django-contact-form==1.4.2
django-contrib-comments==1.7.3
django-overextends==0.4.3
django-redis==4.8.0
django-rq==0.9.6
django-sortedm2m==1.5.0
django-statici18n==1.4.0
django-tastypie==0.14.0
html5lib==1.0b10
idna==2.6
jsonfield==2.0.2
lxml==3.8.0
Markdown==2.6.9
MySQL-python==1.2.5
oauthlib==2.0.4
pathlib==1.0.1
pkg-resources==0.0.0
Pootle==2.9.0rc1
pycountry==17.5.14
python-dateutil==2.6.1
python-Levenshtein==0.12.0
python-mimeparse==1.6.0
python-openid==2.2.5
pytz==2017.2
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
rq==0.8.2
scandir==1.5
six==1.11.0
stemming==1.0.1
translate-toolkit==2.2.5
urllib3==1.22
webassets==0.12.1
webencodings==0.5.1
